refactor: share TUN worker-thread join between platforms
